Stack -> queue and using a single queue
