+++ __tests__/setup/observableToPromise.js
@@ -1,19 +1,24 @@
-import createPromiseStack from './createPromiseStack';
+import createPromiseQueue from './createPromiseQueue';
 
 export default function observableToPromise(observable) {
-  let errors = createPromiseStack();
-  let completes = createPromiseStack();
-  let nexts = createPromiseStack();
+  let queue = createPromiseQueue();
+
   const subscription = observable.subscribe({
-    complete: completes.push,
-    error: errors.push,
-    next: nexts.push
+    complete() {
+      queue.push({ complete: {} });
+      subscription.unsubscribe();
+    },
+    error(error) {
+      queue.push({ error });
+      subscription.unsubscribe();
+    },
+    next(next) {
+      queue.push({ next });
+    }
   });
 
   return {
-    complete: () => completes.pop(),
-    error: () => errors.pop(),
-    next: () => nexts.pop(),
+    shift: queue.shift.bind(queue),
     unsubscribe: subscription.unsubscribe.bind(subscription)
   };
 }
